Rename show() to debug()
Also renames debug.py to debugger.py to avoid import conflicts.
