+++ tests/fixtures/good_definition_files/debug_definition.py
@@ -1,12 +1,12 @@
 # noqa: INP001
-from ehrql import create_dataset
-from ehrql.debug import show, stop
+from ehrql import create_dataset, debug
+from ehrql.debugger import stop
 from ehrql.tables.core import patients
 
 
 dataset = create_dataset()
 dataset.sex = patients.sex
-show("Hello")
+debug("Hello")
 dataset.define_population(patients.date_of_birth.is_on_or_after("2000-01-01"))
 stop()
 dataset.year_of_birth = patients.date_of_birth.year
