+++ tests/unit/test_debug.py
@@ -1,18 +1,19 @@
 import textwrap
 
-from ehrql.debug import show, stop
+from ehrql import debug
+from ehrql.debugger import stop
 from ehrql.query_engines.in_memory_database import PatientColumn
 
 
 def test_show_string(capsys):
     expected_output = textwrap.dedent(
         """
-        Debug line 15:
+        Debug line 16:
         'Hello'
         """
     ).strip()
 
-    show("Hello")
+    debug("Hello")
     captured = capsys.readouterr()
     assert captured.err.strip() == expected_output, captured.err
 
@@ -20,13 +21,13 @@ def test_show_string(capsys):
 def test_show_int_variable(capsys):
     expected_output = textwrap.dedent(
         """
-        Debug line 29:
+        Debug line 30:
         12
         """
     ).strip()
 
     foo = 12
-    show(foo)
+    debug(foo)
     captured = capsys.readouterr()
     assert captured.err.strip() == expected_output, captured.err
 
@@ -34,7 +35,7 @@ def test_show_int_variable(capsys):
 def test_show_multiple_variables(capsys):
     expected_output = textwrap.dedent(
         """
-        Debug line 45:
+        Debug line 46:
         12
         'Hello'
         """
@@ -42,7 +43,7 @@ def test_show_multiple_variables(capsys):
 
     foo = 12
     bar = "Hello"
-    show(foo, bar)
+    debug(foo, bar)
     captured = capsys.readouterr()
     assert captured.err.strip() == expected_output, captured.err
 
@@ -50,12 +51,12 @@ def test_show_multiple_variables(capsys):
 def test_show_with_label(capsys):
     expected_output = textwrap.dedent(
         """
-        Debug line 58: Number
+        Debug line 59: Number
         14
         """
     ).strip()
 
-    show(14, label="Number")
+    debug(14, label="Number")
     captured = capsys.readouterr()
     assert captured.err.strip() == expected_output, captured.err
 
@@ -63,7 +64,7 @@ def test_show_with_label(capsys):
 def test_show_formatted_table(capsys):
     expected_output = textwrap.dedent(
         """
-        Debug line 80:
+        Debug line 81:
         patient_id        | value
         ------------------+------------------
         1                 | 101
@@ -77,7 +78,7 @@ def test_show_formatted_table(capsys):
         2 | 201
         """
     )
-    show(c)
+    debug(c)
     captured = capsys.readouterr()
     assert captured.err.strip() == expected_output, captured.err
 
@@ -85,7 +86,7 @@ def test_show_formatted_table(capsys):
 def test_show_truncated_table(capsys):
     expected_output = textwrap.dedent(
         """
-        Debug line 106:
+        Debug line 107:
         patient_id        | value
         ------------------+------------------
         1                 | 101
@@ -103,7 +104,7 @@ def test_show_truncated_table(capsys):
         """
     )
 
-    show(c, head=1, tail=1)
+    debug(c, head=1, tail=1)
     captured = capsys.readouterr()
     assert captured.err.strip() == expected_output, captured.err
 
@@ -111,10 +112,10 @@ def test_show_truncated_table(capsys):
 def test_stop(capsys):
     stop()
     captured = capsys.readouterr()
-    assert captured.err.strip() == "Stopping at line 112"
+    assert captured.err.strip() == "Stopping at line 113"
 
 
 def test_stop_with_head_and_tail(capsys):
     stop(head=1, tail=1)
     captured = capsys.readouterr()
-    assert captured.err.strip() == "Stopping at line 118"
+    assert captured.err.strip() == "Stopping at line 119"
